Partially corrected client filtering
